+++ app/home/turnofafriendlycard/page.tsx
@@ -1,6 +1,12 @@
+import { Metadata } from "next";
 import { Lobster } from "next/font/google";
 import React from "react";
 const font = Lobster({ subsets: ["latin"], weight: ["400"] });
+
+export const metadata: Metadata = {
+  title: "Guessers | Turn Of A Friendly Card",
+  description: "But the game never ends",
+};
 
 const TurnOfAFriendlyCard = () => {
   return (
