+++ app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { getStats } from "@/api/authentication";
 import Loader from "@/components/common/Loader";
-import LoginOrRegister from "@/components/LoginOrRegister";
+import LoginOrRegister from "@/components/auth/LoginOrRegister";
 import Welcomer from "@/components/Welcomer";
 import { useSession } from "next-auth/react";
 import Image from "next/image";
