+++ components/room/event/pages/EventCreateContent.tsx
@@ -3,8 +3,8 @@ import { createEvent } from "@/api/event";
 import { getRoom } from "@/api/room";
 import PrimaryButton from "@/components/common/button/PrimaryButton";
 import Loader from "@/components/common/Loader";
-import CustomInputField from "@/components/form/CustomInputField";
-import RoomName from "@/components/room/RoomName";
+import CustomInputField from "@/components/common/CustomInputField";
+import RoomName from "@/components/room/layout/RoomName";
 import { IRoomBasic } from "@/types/IRoom.model";
 import { Formik, Form, FieldArray } from "formik";
 import { useRouter } from "next/router";
