+++ components/room/guesspaper/GuessPaperDetails.tsx
