+++ components/room/pages/invite/RoomInviteContent.tsx
@@ -3,12 +3,12 @@ import { getRoom, invitePeople } from "@/api/room";
 import PrimaryButton from "@/components/common/button/PrimaryButton";
 import ComponentTitle from "@/components/common/ComponentTitle";
 import Loader from "@/components/common/Loader";
-import CustomInputField from "@/components/form/CustomInputField";
+import CustomInputField from "@/components/common/CustomInputField";
 import { IRoomBasic } from "@/types/IRoom.model";
 import { Formik, Form } from "formik";
 import React, { useEffect, useState } from "react";
 import { toast } from "react-toastify";
-import RoomName from "../RoomName";
+import RoomName from "../../layout/RoomName";
 
 interface IRoomInviteContentProps {
   params: { roomId: string };
