+++ components/room/pages/ranks/RoomTopPredictors.tsx
